Separated directories with server and client
The Android Client is now in a different directory and all the server
files were added. The server contains the protocol and connection
classes needed to establish the link with the client. The server also
recalculates the relevance of the reports constantly until there are no
recent reports. It also recalculates the color of the Polylines in the
map, but (for some weird reason) the Client's update button does not
repaint the Polylines with the recalculated color.